wip: remove inline rules
